Initial checkin of GTK4 based rewrite

+++ src/log_text_contrast.rs
@@ -1,13 +1,15 @@
 // https://stackoverflow.com/questions/9733288/how-to-programmatically-calculate-the-contrast-ratio-between-two-colors
 
-fn lum_mapper(v: f64) -> f64 {
+use gtk4_helper::gtk::gdk;
+
+fn lum_mapper(v: f32) -> f32 {
     if v <= 0.03928 {
          return v / 12.92;
     }
-    return f64::powf( (v + 0.055) / 1.055, 2.4 );
+    return f32::powf( (v + 0.055) / 1.055, 2.4 );
 }
 
-fn luminance(r: f64, g: f64, b: f64) -> f64 {
+fn luminance(r: f32, g: f32, b: f32) -> f32 {
     let r = lum_mapper(r);
     let g = lum_mapper(g);
     let b = lum_mapper(b);
@@ -15,21 +17,21 @@ fn luminance(r: f64, g: f64, b: f64) -> f64 {
     return r * 0.2126 + g * 0.7152 + b * 0.0722;
 }
 
-fn maxf(a: f64, b: f64) -> f64 {
+fn maxf(a: f32, b: f32) -> f32 {
     if a > b {
         return a;
     }
     return b;
 }
 
-fn minf(a: f64, b: f64) -> f64 {
+fn minf(a: f32, b: f32) -> f32 {
     if a < b {
         return a;
     }
     return b;
 }
 
-fn contrast(col1: &gdk::RGBA, col2: &gdk::RGBA) -> f64 {
+fn contrast(col1: &gdk::RGBA, col2: &gdk::RGBA) -> f32 {
     let lum1 = luminance(col1.red, col1.green, col1.blue);
     let lum2 = luminance(col2.red, col2.green, col2.blue);
     let brightest = maxf(lum1, lum2);
